+++ src/quacc/recipes/dftb/core.py
@@ -7,10 +7,10 @@ from ase.calculators.dftb import Dftb
 
 from quacc import job
 from quacc.schemas.ase import summarize_run
-from quacc.schemas.atoms import fetch_atoms
-from quacc.util.calc import run_calc
-from quacc.util.dicts import merge_dicts
-from quacc.util.files import check_logfile
+from quacc.utils.calc import run_calc
+from quacc.utils.dicts import merge_dicts
+from quacc.utils.files import check_logfile
+from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms