+++ tests/utils/test_db.py
@@ -6,7 +6,7 @@ from maggma.stores import MemoryStore
 
 from quacc import SETTINGS
 from quacc.recipes.emt.core import static_job
-from quacc.util.db import results_to_db
+from quacc.utils.db import results_to_db
 
 try:
     import covalent as ct
@@ -20,7 +20,7 @@ except ImportError:
     reason="This test is only meant to be run on GitHub Actions with Covalent",
 )
 def test_covalent_to_db():
-    from quacc.util.db import covalent_to_db
+    from quacc.utils.db import covalent_to_db
 
     store = MemoryStore(collection_name="db1")
     covalent_to_db(store)